+++ Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,5 +1,6 @@
-﻿using AutoMapper;
+﻿using Domain.Constants;
 using Domain.Exceptions;
+using Domain.Interfaces;
 using Domain.Repositories;
 using MediatR;
 using Microsoft.Extensions.Logging;
@@ -8,14 +9,21 @@ namespace Application.Restaurants.Commands.DeleteRestaurant;
 
 public class DeleteRestaurantCommandHandler(
     ILogger<DeleteRestaurantCommand> logger,
-    IMapper mapper,
-    IRestaurantsRepository restaurantsRepository) : IRequestHandler<DeleteRestaurantCommand>
+    IRestaurantsRepository restaurantsRepository,
+    IRestaurantAuthorizationService restaurantAuthorizationService
+) : IRequestHandler<DeleteRestaurantCommand>
 {
     public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
     {
         logger.LogInformation("Deleting restaurant with id {RestaurantId}", request.Id);
         var restaurant = await restaurantsRepository.GetRestaurantById(request.Id);
         if (restaurant is null) throw new NotFoundException(nameof(restaurant), request.Id.ToString());
+
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperations.Delete))
+        {
+            throw  new ForbiddenException();
+        }
+
         await restaurantsRepository.DeleteRestaurant(restaurant);
     }
 }
